refactor(gazelle): move common code into gazelle/common (#2458)
GitOrigin-RevId: f3c4a968173032d774392372665be94c1c4d2f6f
