+++ gazelle/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
+
+bzl_library(
+    name = "gazelle",
+    srcs = ["gazelle.bzl"],
+    visibility = ["//visibility:public"],
+    deps = ["@bazel_gazelle//:def"],
+)
