+++ gazelle/common/git/gitignore_test.go
@@ -1,4 +1,4 @@
-package gazelle
+package git
 
 import (
 	"fmt"
