+++ gazelle/js/parser/treesitter/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     importpath = "aspect.build/cli/gazelle/js/parser/treesitter",
     visibility = ["//visibility:public"],
     deps = [
+        "//gazelle/common/treesitter",
         "//gazelle/js/parser",
         "@com_github_sirupsen_logrus//:logrus",
         "@com_github_smacker_go_tree_sitter//:go-tree-sitter",
